GUI-VersionForPyCharm/requirements.txt
pyautogui
pygithub
